src: check the return value from `_libssh2_bn_*()` functions (#1354)
Found by oss-fuzz. In `diffie_hellman_sha_algo()`, we were calling
`_libssh2_bn_from_bin()` with data recieved by the server without
checking whether that data was zero-length or ridiculously long.
In the OpenSSL backend, this would cause `_libssh2_bn_from_bin()`
to fail an allocation, which would eventually lead to a NULL
dereference when the bignum was used.

Add the same check for `_libssh2_bn_set_word()` and
`_libssh2_bn_to_bin()`.

+++ src/os400qc3.c
@@ -458,9 +458,9 @@ _libssh2_bn_bits(_libssh2_bn *bn)
 }
 
 int
-_libssh2_bn_from_bin(_libssh2_bn *bn, int len, const unsigned char *val)
+_libssh2_bn_from_bin(_libssh2_bn *bn, size_t len, const unsigned char *val)
 {
-    int i;
+    size_t i;
 
     if(!bn || (len && !val))
         return -1;
